wsgi_app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
beautifulsoup4==4.11.2
bidict==0.22.1
blinker==1.5
Bootstrap-Flask==2.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
cssmin==0.2.0
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
filelock==3.9.0
Flask==2.2.3
Flask-Assets==2.0
Flask-Cors==3.0.10
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
gevent==22.10.2
gevent-websocket==0.10.1
google==3.0.0
google-auth==2.16.1
googlemaps==4.10.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
libsass==0.22.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
onetimepass==1.0.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyotp==2.8.0
PyPDF2==3.0.1
pypiwin32==223
PyQRCode==1.2.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyScss==1.4.0
pystray==0.19.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7.1
pywin32==305
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
URLObject==2.4.3
virtualenv==20.19.0
virtualenv-clone==0.5.7
Wand==0.6.11
webassets==2.0
websocket-client==1.5.1
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2

pdf2image~=1.16.3